Refactor Main file to add helpers and modules for future user stories
