+++ src/modules/list_upcoming_dates.py
@@ -1,0 +1,5 @@
+import datetime
+
+
+def list_upcoming_birthdays(individuals):
+    pass
